% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarystats.R
\name{calc_summary_stats}
\alias{calc_summary_stats}
\title{Calculate Summary Statistics}
\usage{
calc_summary_stats(data, grouping_variable, data_variable)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{grouping_variable}{A character string specifying the variable by which the data
should be grouped, therefore named grouping_variable.}

\item{data_variable}{A character string specifying the variable for which summary
statistics will be calculated. The variable is wthin a dataset, thus named data_variable.}
}
\value{
A data frame containing the summary statistics range, mean, median,
and standard deviation.
}
\description{
'calc_summary_stats' calculates summary statistics for a given variable within a dataset,
grouped by another variable.
}
\examples{
sepal_length_summary <- calc_summary_stats(data = iris, grouping_variable = Species, data_variable = Sepal.Length)

}
